requirements.txt
fastapi
uvicorn
playwright
websockets
pydantic
python-dotenv
opencv-python
pytesseract
Pillow